# docs/openapi.json
{
  "components": {
    "schemas": {
      "AuditEntryResponse": {
        "description": "One entry from an entity's change history.",
        "properties": {
          "action": {
            "description": "`create`, `update` or `delete`.",
            "title": "Action",
            "type": "string"
          },
          "actor_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Their name, or null if the account was deleted. The entry survives it.",
            "title": "Actor Name"
          },
          "actor_user_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Who made the change.",
            "title": "Actor User Id"
          },
          "after": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "State after the change.",
            "title": "After"
          },
          "at": {
            "description": "ISO-8601 UTC.",
            "title": "At",
            "type": "string"
          },
          "before": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "State before the change.",
            "title": "Before"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "action",
          "at"
        ],
        "title": "AuditEntryResponse",
        "type": "object"
      },
...
      "GradeCreateRequest": {
        "description": "A new grade.",
        "properties": {
          "course_id": {
            "examples": [
              "CS101"
            ],
            "minLength": 1,
            "title": "Course Id",
            "type": "string"
          },
          "date": {
            "description": "ISO `YYYY-MM-DD`. `DD-MM-YYYY`, `DD.MM.YYYY` and `DD/MM/YYYY` are also accepted and normalised.",
            "examples": [
              "2026-01-15"
            ],
            "title": "Date",
            "type": "string"
          },
          "notes": {
            "default": "",
            "maxLength": 2000,
            "title": "Notes",
            "type": "string"
          },
          "score": {
            "description": "Must not exceed the course maximum.",
            "examples": [
              85.0
            ],
            "minimum": 0.0,
            "title": "Score",
            "type": "number"
          },
          "student_id": {
            "examples": [
              "S001"
            ],
            "minLength": 1,
            "title": "Student Id",
            "type": "string"
          },
          "title": {
            "default": "",
            "examples": [
              "Midterm"
            ],
            "maxLength": 100,
            "title": "Title",
            "type": "string"
          },
          "weight": {
            "default": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "student_id",
          "course_id",
          "score",
          "date"
        ],
        "title": "GradeCreateRequest",
        "type": "object"
      },
      "GradeResponse": {
        "description": "A recorded grade.\n\n`percentage` and `is_passing` are computed server-side deliberately: two clients\nderiving them independently is two chances to disagree with the report the same\nnumbers appear in.",
        "properties": {
          "course_id": {
            "examples": [
              "CS101"
            ],
            "title": "Course Id",
            "type": "string"
          },
          "course_name": {
            "examples": [
              "Intro to Programming"
            ],
            "title": "Course Name",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "date": {
            "description": "Award date, ISO `YYYY-MM-DD`.",
            "examples": [
              "2026-01-15"
            ],
            "title": "Date",
            "type": "string"
          },
          "grade_id": {
            "description": "Database identifier.",
            "title": "Grade Id",
            "type": "integer"
          },
          "graded_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Who recorded it.",
            "title": "Graded By"
          },
          "is_passing": {
            "description": "Whether the score reaches the course threshold.",
            "title": "Is Passing",
            "type": "boolean"
          },
          "max_grade": {
            "description": "The course maximum.",
            "examples": [
              100.0
            ],
            "title": "Max Grade",
            "type": "number"
          },
          "notes": {
            "description": "Free-text remark.",
            "title": "Notes",
            "type": "string"
          },
          "percentage": {
            "description": "Score as a percentage of the maximum.",
            "examples": [
              85.0
            ],
            "title": "Percentage",
            "type": "number"
          },
          "score": {
            "description": "Points awarded.",
            "examples": [
              85.0
            ],
            "title": "Score",
            "type": "number"
          },
          "student_id": {
            "examples": [
              "S001"
            ],
            "title": "Student Id",
            "type": "string"
          },
          "student_name": {
            "examples": [
              "Anna Schmidt"
            ],
            "title": "Student Name",
            "type": "string"
          },
          "title": {
            "description": "Assessment name, e.g. `Midterm`. Empty for a single mark.",
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "weight": {
            "description": "Relative weight in the course average.",
            "examples": [
              1.0
            ],
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "grade_id",
          "student_id",
          "student_name",
          "course_id",
          "course_name",
          "title",
          "score",
          "max_grade",
          "percentage",
          "is_passing",
          "weight",
          "date",
          "notes"
        ],
        "title": "GradeResponse",
        "type": "object"
      },
      "GradeUpdateRequest": {
        "description": "Changes to a grade. Omitted fields are left alone.",
        "properties": {
          "date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "score": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "weight": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight"
          }
        },
        "title": "GradeUpdateRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PageResponse_GradeResponse_": {
        "properties": {
          "items": {
            "description": "The rows on this page.",
            "items": {
              "$ref": "#/components/schemas/GradeResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "description": "The 1-based page number returned.",
            "title": "Page",
            "type": "integer"
          },
          "pages": {
            "description": "Total number of pages.",
            "title": "Pages",
            "type": "integer"
          },
          "size": {
            "description": "Rows per page.",
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "description": "Rows matching the query overall, ignoring paging.",
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "PageResponse[GradeResponse]",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "\nStudent grade management: students, courses, enrolments, grades, reports and an\nAI assistant.\n\n### Errors\n\nEvery failure returns `application/problem+json` (RFC 9457) carrying a stable\nmachine-readable `code` and a structured `context`:\n\n```json\n{ \"type\": \"about:blank#STUDENT_NOT_FOUND\", \"status\": 404,\n  \"code\": \"STUDENT_NOT_FOUND\", \"detail\": \"No student with id 'S999'.\",\n  \"context\": { \"student_id\": \"S999\" } }\n```\n\n**Clients should render the `code`, not the `detail`.** `detail` is a\ndeveloper-facing string and is only ever English; the interface translates the code\ninto the reader's language. This is why the API ships no message catalogue.\n\n### Authorization\n\nAuthentication is a session cookie (`HttpOnly`, `SameSite=Lax`). Access is filtered\nin the query rather than checked in the handler, so what you can see depends on your\nrole:\n\n| Role | Students | Courses | Grades |\n|---|---|---|---|\n| `student` | only themselves | those they are enrolled on | only their own |\n| `teacher` | those enrolled on a course they own | those they own | within their own courses |\n| `admin` | all | all | all |\n| `superadmin` | all, plus AI provider configuration | all | all |\n\nA list endpoint returns what you may see, never a `403`. Only *actions* beyond your\nrole return `403`.\n",
    "title": "Grade Tracker API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/grades": {
      "get": {
        "description": "Paginated, sortable and searchable, filtered to what your role permits.\n\nA grade is visible only when **both** its student and its course are. That is why a teacher cannot read a shared student's marks from a colleague's course, even though the student is otherwise in their scope.",
        "operationId": "list_grades_grades_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 200,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "description": "`date`, `score`, `student`, `course`, `title` or `created`; `-` reverses.",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`date`, `score`, `student`, `course`, `title` or `created`; `-` reverses.",
              "title": "Sort"
            }
          },
          {
            "description": "Free text over student, course and title.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free text over student, course and title.",
              "title": "Q"
            }
          },
          {
            "description": "Only this student's grades.",
            "in": "query",
            "name": "student_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only this student's grades.",
              "title": "Student Id"
            }
          },
          {
            "description": "Only this course's grades.",
            "in": "query",
            "name": "course_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only this course's grades.",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse_GradeResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List grades",
        "tags": [
          "Grades"
        ]
      },
      "post": {
        "description": "The date accepts ISO `YYYY-MM-DD` as well as `DD-MM-YYYY`, `DD.MM.YYYY` and `DD/MM/YYYY`, and is normalised to ISO on the way in.\n\nUse `title` and `weight` for several assessments in one course — a midterm and a final with different weights.",
        "operationId": "record_grade_grades_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GradeCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GradeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "`FORBIDDEN` — you cannot grade this course."
          },
          "404": {
            "description": "`STUDENT_NOT_FOUND` or `COURSE_NOT_FOUND`."
          },
          "422": {
            "description": "`VALIDATION_ERROR` — score out of range, or an unparseable date."
          }
        },
        "summary": "Record a grade",
        "tags": [
          "Grades"
        ]
      }
    },
    "/grades/{grade_id}": {
      "delete": {
        "description": "Soft delete. The row is retained and excluded from reads, because an altered mark is exactly the kind of change a student may later dispute.",
        "operationId": "retire_grade_grades__grade_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "grade_id",
            "required": true,
            "schema": {
              "title": "Grade Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "`GRADE_NOT_FOUND`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retire a grade",
        "tags": [
          "Grades"
        ]
      },
      "get": {
        "description": "Fetch one grade.",
        "operationId": "get_grade_grades__grade_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "grade_id",
            "required": true,
            "schema": {
              "title": "Grade Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GradeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "`GRADE_NOT_FOUND` — unknown, or outside your scope."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fetch one grade",
        "tags": [
          "Grades"
        ]
      },
      "patch": {
        "description": "The previous values are written to the audit trail.",
        "operationId": "amend_grade_grades__grade_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "grade_id",
            "required": true,
            "schema": {
              "title": "Grade Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GradeUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GradeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "description": "`FORBIDDEN` — you cannot grade this course."
          },
          "404": {
            "description": "`GRADE_NOT_FOUND`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Amend a grade",
        "tags": [
          "Grades"
        ]
      }
    },
    "/grades/{grade_id}/history": {
      "get": {
        "description": "Who changed this mark, when, and from what. Scope is checked first, so the trail cannot be used to read a grade indirectly.",
        "operationId": "grade_history_grades__grade_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "grade_id",
            "required": true,
            "schema": {
              "title": "Grade Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEntryResponse"
                  },
                  "title": "Response Grade History Grades  Grade Id  History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "`GRADE_NOT_FOUND`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "A grade's change history",
        "tags": [
          "Grades"
        ]
      }
    },
...
  },
...
}